back_end/requirements.txt
albucore==0.0.24
albumentations==2.0.8
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
doclayout_yolo==0.0.4
dotenv==0.9.9
easyocr==1.7.2
faiss-cpu==1.11.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
fonttools==4.58.4
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.4
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.0
openai==1.86.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
owlready2==0.48
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prompt_toolkit==3.0.51
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.1
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
redis==6.2.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
setuptools==78.1.1
shapely==2.1.1
simple-websocket==1.1.0
simsimd==6.4.9
six==1.17.0
sniffio==1.3.1
stringzilla==3.12.5
sympy==1.14.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
